Split tets even further
